// src/types.ts
export type RawAddressInfo = {
  'prefecture_jis_code': string;
  'city_jis_code': string;
  'zip_code': string;
  'prefecture_name_kana': string;
  'city_name_kana': string;
  'town_name_kana': string;
  'prefecture_name': string;
  'city_name': string;
  'town_name': string;
}

export type AddressInfo = {
  prefectureJisCode: string;
  cityJisCode: string;
  zipCode: string;
  prefectureNameKana: string;
  cityNameKana: string;
  townNameKana: string;
  prefectureName: string;
  cityName: string;
  townName: string;
}